introduce image panel component
