+++ src/components/ImageView/ImagePanel/ImagePanelComponent.scss
@@ -1,0 +1,17 @@
+.image-view-div {
+    width: 100%;
+    height: 100%;
+    position: relative;
+    user-select: none;
+
+    .tag-image-ratio {
+        transition: opacity 0.5s;
+        position: absolute;
+        z-index: 5;
+        pointer-events: none;
+
+        .bp3-tag {
+            left: -50%;
+        }
+    }
+}
